fixed streamout problem with displacement

+++ Include/instructions/isjkj.hh
@@ -7,7 +7,7 @@ class isjkj : public Fjk
 	{
 	    stringstream ss;
 	    ss << setfill('0') << setw(16) << hex << PROC[me()].X(_j).i() <<
-                " " << setw(1) << _k;
+                " " << setw(1) << u32(_k);
 	    PROC[me()].X(_j).i() = PROC[me()].X(_j).i() + _k;
 	    ss << " " << setw(16) << PROC[me()].X(_j).i() << dec << setfill(' ');
 	    _trace = ss.str();
